+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## Master
+
+* Make Napybara::Element#inspect more readable.
+* Be able to find the root of an element.
+
 ## 0.6.0
 
 Allow Capybara options to be passed to finders.